backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi>=0.111.0
uvicorn>=0.29.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
httpx>=0.27.0
requests>=2.31.0
python-dotenv>=1.0.0
playwright>=1.42.0
celery[redis]>=5.4.0
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Vision & AI Models
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.40.0
pillow>=10.2.0
